Create separate service
