+++ src/main/java/org/folio/qm/service/LinkingRulesService.java
@@ -1,0 +1,9 @@
+package org.folio.qm.service;
+
+import java.util.List;
+import org.folio.qm.client.LinksClient;
+
+public interface LinkingRulesService {
+
+  List<LinksClient.LinkingRuleDto> getLinkingRules();
+}
