+++ src/main/java/org/folio/qm/service/LinksService.java
@@ -1,7 +1,5 @@
 package org.folio.qm.service;
 
-import java.util.List;
-import org.folio.qm.client.LinksClient;
 import org.folio.qm.domain.dto.QuickMarc;
 
 public interface LinksService {
@@ -9,6 +7,4 @@ public interface LinksService {
   void setRecordLinks(QuickMarc qmRecord);
 
   void updateRecordLinks(QuickMarc qmRecord);
-
-  List<LinksClient.LinkingRuleDto> getLinkingRules();
 }
